Migrate Dialog component to TypeScript
